difftest: instantiate Delayer inside endpiont
With Hierarchical Wiring (require Chisel6 and difftestConfig H),
we can collect signals from DUT to Difftest, like SimTop.xx.yy.

Previous we instantiate Delayer before addSource, the Delayer is
instantiated inside DUT. This change move it after addSource, so
Design and Verification code can be seperated when needEndpoint and
Hierarchical Wiring.

When Hierarchical wiring, We use Endpoint by default.
